Finally fix https://github.com/kristapsdz/kcgi/issues/51
This is due to bad documentation in rights(4), which states

"When a file descriptor is created by a function such as accept(2),
     ...  it is assigned all capability rights."

But this is untrue, as the socket created with the accept(2) fails with
limited capabilities.  It seems, instead, to inherit the (some?) rights
from the descriptor from which it was created.