Add renamed khttp_urlpart and family, which in form is like kutil_urlpart except without the "struct kreq".  It also adds khttp_vurlpart et al., which exposes the va_list interface.
The new functions have two small behaviour difference: if the MIME suffix
is an empty string, the dot-suffix is not printed; and if the page is NULL,
it is simply not printed as if it were an empty string.  Furthermore, an
empty or NULL page does not print the MIME suffix too.  This makes output
conform to URL standards.

The old functions are left with the existing behaviour *except* that the
MIME suffix is not printed if the suffix is an empty string.