When using kutil_openlog(), a file may be opened with freopen().  The first log message via kutil_log() et al. will thus emit a fstat() and possibly an lseek() to find the end of the file before appending.  This was breaking the sandbox filter, which has write allowed but not fstat/lseek.
While here, remove the file size limitation for the same reason.

This might still be broken on systems that use alternatives to fstat like the
weird Linux fstat64 or whatever.  There's no way to test this, unfortunately,
because the seccomp debugging will print the offending system call to... stderr.
Which is what can't be written to.  Linux...