Abandoned seccomp by default.
seccomp is causing too many problems, and I don't have the resources to
test all the hardware and software combinations that are required.

I'll continue to update the machinery when patches come in, but until
Linux has a security mechanism that doesn't require a corporate
infrastructure to operate, sandboxing for this operating system is
disabled by default.

If you're running kcgi in a security-sensitive environment, seccomp can
be enabled in the Makefile (instructions are therein).  If you're
packaging kcgi for seccomp systems, please don't enable it unless you've
tested for the target platform!

I strongly urge security-sensitive applications to move to OpenBSD or
FreeBSD, whose security mechanisms are much more friendly to
development.